requirements.txt
asgiref
django
django-cors-headers
django-filter
djangorestframework
markdown
sqlparse
python-dotenv

gunicorn
dj-database-url
psycopg2-binary

# Comando de inicio: gunicorn config.wsgi:application
